Several changes, refactoring and imrpoving readability

+++ src/main/java/com/cloudcoders/gestaca/persistance/TaughtCourseDAOImpl.java
@@ -64,7 +64,7 @@ public class TaughtCourseDAOImpl implements ITaughtCourseDAO {
   }
 
   @Override
-  public TaughtCourse get(int id) {
+  public TaughtCourse get(long id) {
     return null;
   }
